- Functional for rustscan discovery and detailed scans (top 1000 vs top 65535)
