+++ Plugins/CoreAPI/client/pages/project-view/project-view-dashboard.jsx
@@ -9,6 +9,8 @@ const useStyles = makeStyles((theme) => ({}));
 const ProjectViewDashboard = ({ project }) => {
   const classes = useStyles();
 
+  console.log(project);
+
   return (
     <Grid container spacing={3}>
       <Grid item xl={4} lg={4} sm={6} xs={12}>
