--- Plugins/CoreAPI/client/pages/projects-list/views-timeline-view.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import { Components, registerComponent } from "@penpal/core";
-import CalendarViewDayIcon from "@mui/icons-material/CalendarViewDay";
-
-const ProjectsViewTimelineView = ({ projects }) => {
-  return null;
-};
-
-export const Icon = <CalendarViewDayIcon />;
-export const Name = "Timeline View";
-
-registerComponent("ProjectsViewTimelineView", ProjectsViewTimelineView);
