+++ docker-compose.common.yaml
@@ -31,7 +31,7 @@ services:
       - "3001:3001"
     environment:
       - RUN_LOCATION=${RUN_LOCATION}
-      - OFFLINE=true
+      - OFFLINE=false
     command: npm run server
     networks:
       - penpal
